First stab at scene saving
