+++ SRPCommon/SRPCommon.csproj
@@ -111,6 +111,7 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Scene\Scene_Serialisation.cs" />
     <Compile Include="UserProperties\IMatrixProperty.cs" />
     <Compile Include="UserProperties\IPropertySource.cs" />
     <Compile Include="Interfaces\IWorkspace.cs" />
